+++ src/pages/SignInPage.tsx
@@ -1,17 +1,28 @@
-import { TextField, Button } from "@mui/material"
+import { TextField, Button } from '@mui/material'
 
 const SignInPage = () => {
-
     return (
         <div>
             <p>Kirjaudu sisään</p>
             <div>
-                <TextField style={{marginRight: 5}}id="outlined-basic" label="Käyttäjätunnus" variant="outlined" />
-                <TextField style={{marginLeft: 5}}id="outlined-basic" label="Salasana" variant="outlined" />
+                <TextField
+                    style={{ marginRight: 5 }}
+                    id="outlined-basic"
+                    label="Käyttäjätunnus"
+                    variant="outlined"
+                />
+                <TextField
+                    style={{ marginLeft: 5 }}
+                    id="outlined-basic"
+                    label="Salasana"
+                    variant="outlined"
+                />
             </div>
-            <Button style={{marginTop: 20}}variant="contained">Kirjaudu Sisään</Button>
+            <Button style={{ marginTop: 20 }} variant="contained">
+                Kirjaudu Sisään
+            </Button>
         </div>
     )
-    }
-    
-    export default SignInPage
+}
+
+export default SignInPage
